fix dtype conversion issue for numpy array

+++ src/pcgrad.py
@@ -115,7 +115,7 @@ class SimPGrad:
             loss_dict[loss_name].backward(retain_graph=True)            
             for name, p, prev_grad, is_reset in zip(names, params, prev_grads, reset_flags):
                 if p.grad is not None:            
-                    grad_array = p.grad.detach().cpu().numpy()
+                    grad_array = p.grad.detach().cpu().to(torch.float16).numpy()
                     p.grad, curr_g_norm, prev_g_norm, cosim = self._project_non_conflict_info(p.grad, prev_grad)
                     self._update_info(name, prev_g_norm, curr_g_norm, cosim, loss_name, is_reset, grad_array)
                     
@@ -138,7 +138,7 @@ class SimPGrad:
             loss_dict[loss_name].backward(retain_graph=True)            
             for name, p, prev_grad, is_reset in zip(names, params, prev_grads, reset_flags):
                 if p.grad is not None:            
-                    grad_array = p.grad.detach().cpu().numpy()
+                    grad_array = p.grad.detach().cpu().to(torch.float16).numpy()
                     p.grad, curr_g_norm, prev_g_norm, cosim = self._add_grad_info(p.grad, prev_grad)
                     self._update_info(name, prev_g_norm, curr_g_norm, cosim, loss_name, is_reset, grad_array)
